refactor(functions): in light of app package refactor
